- Update CollatzSequence calculation to use long instead of int - Problem 14 solved

+++ euler/src/test/java/com/ijemmy/euler/CollatzSequenceTest.java
@@ -42,7 +42,7 @@ public class CollatzSequenceTest {
 	
 	@Test
 	public void getSeqLengthUseCache() {
-		CollatzSequence.seqLengthMap = new HashMap<Integer, Integer>();
+		CollatzSequence.seqLengthMap = new HashMap<Long, Integer>();
 		CollatzSequence.getLength(13);
 		assertEquals(10 , CollatzSequence.seqLengthMap.size());
 	}
